requirements.txt
geopandas==1.0.1
openai==1.58.1
pandas==2.2.3
streamlit_folium==0.24.0
pyshp==2.3.1
matplotlib==3.10.0
groq==0.13.1
plotly==5.24.1
numpy==2.2.0
streamlit-chat==0.1.1
streamlit-extr==0.5.0